requirements.txt
Django>=4.0.4,<4.1.0
djangorestframework>=3.13.1,<3.14.0
psycopg2>=2.9.3,<3.0.0
